uk-esos-app-api/src/main/java/uk/gov/esos/api/web/controller/authorization/RoleController.java
package uk.gov.esos.api.web.controller.authorization;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import uk.gov.esos.api.authorization.core.domain.dto.RoleDTO;
import uk.gov.esos.api.authorization.core.service.RoleService;
import uk.gov.esos.api.authorization.regulator.domain.RegulatorRolePermissionsDTO;
import uk.gov.esos.api.authorization.regulator.service.RegulatorRoleService;
import uk.gov.esos.api.web.constants.SwaggerApiInfo;
import uk.gov.esos.api.web.controller.exception.ErrorResponse;
import uk.gov.esos.api.web.security.Authorized;

import java.util.List;

import static uk.gov.esos.api.web.constants.SwaggerApiInfo.INTERNAL_SERVER_ERROR;
import static uk.gov.esos.api.web.constants.SwaggerApiInfo.OK;

@RestController
@RequestMapping(path = "/v1.0/authorities")
@Tag(name = "Authorities")
@RequiredArgsConstructor
public class RoleController {

    private final RoleService roleService;
    private final RegulatorRoleService regulatorRoleService;

    /**
     * Returns all operator roles.
     *
     * @return List of {@link RoleDTO}
     */
    @GetMapping(path = "/account/{accountId}/operator-role-codes")
    @Operation(summary = "Retrieves the operator roles")
    @ApiResponse(responseCode = "200", description = OK, content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, array = @ArraySchema(schema = @Schema(implementation = RoleDTO.class))))
    @ApiResponse(responseCode = "403", description = SwaggerApiInfo.FORBIDDEN, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @ApiResponse(responseCode = "500", description = INTERNAL_SERVER_ERROR, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @Authorized(resourceId = "#accountId")
    public ResponseEntity<List<RoleDTO>> getOperatorRoleCodes(
            @PathVariable("accountId") @Parameter(description = "The account id") Long accountId) {
        List<RoleDTO> roles = roleService.getOperatorRoles();
        return new ResponseEntity<>(roles, HttpStatus.OK);
    }

    /**
     * Returns all regulator roles.
     * @return List of {@link RegulatorRolePermissionsDTO}
     */
    @GetMapping(path = "/regulator-roles")
    @Operation(summary = "Returns all regulator roles")
    @ApiResponse(responseCode = "200", description = OK, content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, array = @ArraySchema(schema = @Schema(implementation = RegulatorRolePermissionsDTO.class))))
    @ApiResponse(responseCode = "403", description = SwaggerApiInfo.FORBIDDEN, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @ApiResponse(responseCode = "500", description = INTERNAL_SERVER_ERROR, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @Authorized
    public ResponseEntity<List<RegulatorRolePermissionsDTO>> getRegulatorRoles() {
        return new ResponseEntity<>(regulatorRoleService.getRegulatorRoles(), HttpStatus.OK);
    }

    @GetMapping(path = "/verifier-role-codes")
    @Operation(summary = "Retrieves the verifier role codes")
    @ApiResponse(responseCode = "200", description = OK, content = @Content(mediaType = MediaType.APPLICATION_JSON_VALUE, array = @ArraySchema(schema = @Schema(implementation = RoleDTO.class))))
    @ApiResponse(responseCode = "403", description = SwaggerApiInfo.FORBIDDEN, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @ApiResponse(responseCode = "500", description = INTERNAL_SERVER_ERROR, content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE, schema = @Schema(implementation = ErrorResponse.class))})
    @Authorized
    public ResponseEntity<List<RoleDTO>> getVerifierRoleCodes() {
        return new ResponseEntity<>(roleService.getVerifierRoleCodes(), HttpStatus.OK);
    }

}
